p.haul: criu gui integration
After all the code from https://github.com/ThomasJClark/criu-gui-demo
has been integrated into p.haul and squashed into 1 commit in the
previous commit, it is now integrated into p.haul. In addition to the
p.haul integration criu-gui-demo has been rewritten to use flask instead
of webpy. Especially as it offers better handling of where to place the
static content.

The basic usage looks like this:

 host01: ./p.haul-wrap service --web-gui
 host02: ./p.haul-wrap service --web-gui --web-partner host01

If p.haul is started like this a web-interface on both hosts on port
8080 should be available:

 http://host01:8080/

It it now possible to drag and drop a process from one host to another
which will trigger a p.haul controlled process migration in the
background.

Right now the p.haul web-gui only supports migration of the type PID.
There is no support to migrate containers of the type LXC or VZ. If this
is desired a new interface to list those running containers instead of
processes would be required.

Signed-off-by: Adrian Reber <areber@redhat.com>